+++ packages/main.ts
@@ -2,8 +2,11 @@
 // (runtime-only or standalone) has been set in webpack.base.conf with an alias.
 import Vue from 'vue'
 import App from './App.vue'
+import Toast from './toast/index.js';
 
-Vue.config.productionTip = false
+Vue.config.productionTip = false;
+
+Vue.use(Toast);
 
 /* eslint-disable no-new */
 new Vue({